+++ libplateau/CSharpPLATEAU/Geometries/CoordinateReferenceFactory.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 94f5a04f9f0a2224a81299e988177bfe
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
